+++ test/react-native/features/fixtures/app/index.js
@@ -2,24 +2,27 @@ import { AppRegistry, SafeAreaView, StyleSheet, Text } from 'react-native'
 import React, { useEffect, useState } from 'react'
 import { name as appName } from './app.json';
 import { launchScenario, launchFromStartupConfig, ScenarioContext, ScenarioComponent } from '@bugsnag/react-native-performance-scenarios'
-import BugsnagPerformance from '@bugsnag/react-native-performance';
 
-const isStartupTest = launchFromStartupConfig()
+const startupConfig = launchFromStartupConfig()
+
+const initialScenario = startupConfig?.scenario ? { name: startupConfig.scenario } : null
 
 const App = () => {
 
-  const [currentScenario, setCurrentScenario] = useState(null)
+  const [currentScenario, setCurrentScenario] = useState(initialScenario)
 
   useEffect(() => {
-    if (!isStartupTest) launchScenario(setCurrentScenario)
+    if (!startupConfig) {
+      launchScenario(setCurrentScenario)
+    }
   }, [])
 
   return (
     <ScenarioContext.Provider value={ currentScenario }>
       <SafeAreaView style={styles.container}>
-          <Text accessibilityLabel='app-component' testID='app-component'>React Native Performance Test App</Text>
-          <ScenarioComponent />
-        </SafeAreaView>
+        <Text accessibilityLabel='app-component' testID='app-component'>React Native Performance Test App</Text>
+        <ScenarioComponent />
+      </SafeAreaView>
     </ScenarioContext.Provider>
   )
 }