+++ test/react-native/features/fixtures/scenario-launcher/scenarios/core/index.js
@@ -1,4 +1,7 @@
 export * as AppStartScenario from './AppStartScenario'
+export * as AutomaticCustomAppStartScenario from './AutomaticCustomAppStartScenario'
+export * as ManualCustomAppStartScenario from './ManualCustomAppStartScenario'
+export * as AppStartSpanControlScenario from './AppStartSpanControlScenario'
 export * as WithInstrumentedAppStartsScenario from './WithInstrumentedAppStartsScenario'
 export * as BackgroundSpanScenario from './BackgroundSpanScenario'
 export * as ErrorCorrelationScenario from './ErrorCorrelationScenario'